# api/openapi.yaml
openapi: 3.0.3
info:
  title: Stalker API
  description: |
    The API specification developed by GruppOne for the Stalker project.
  contact:
    name: GruppOne
    email: gruppone.swe@gmail.com
    url: 'https://github.com/GruppOne'
  license:
    name: GNU GPLv3
    url: 'https://www.gnu.org/licenses/gpl-3.0-standalone.html'
  version: 0.8.1
servers:
  - url: 'http://localhost:11111'
    description: Development server.
  - url: 'https://{environment}.gruppone.tech/v{majorVersion}'
    variables:
      environment:
        enum:
          - api.staging
          - api
        default: api
      majorVersion:
        default: '0'
    description: Staging and Production servers.
...
paths:
...
  '/organization/{organizationId}/user/{userId}/role':
    delete:
      description: Delete an existing role for a user (without deleting the user or their connection to the organization).
      operationId: delete-organization-organizationId-user-role
      responses:
        '204':
          description: No Content
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/DefaultResponse'
      summary: Delete a role for a user.
      tags:
        - web app
    parameters:
      - $ref: '#/components/parameters/organizationId'
      - $ref: '#/components/parameters/userId'
    post:
      description: Add a new role for a user.
      operationId: post-organization-organizationId-user-role
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                newRole:
                  $ref: ./model/AdministratorType.v1.yaml
              required:
                - newRole
      responses:
        '201':
          description: Created
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          $ref: '#/components/responses/DefaultResponse'
      summary: Create a role for a user in an organization.
      tags:
        - web app
    put:
      summary: Update a role for a user in an organization.
      operationId: put-organization-organizationId-user-userId-role
      responses:
        '204':
          description: No Content
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/DefaultResponse'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                modifiedRole:
                  $ref: ./model/AdministratorType.v1.yaml
              required:
                - modifiedRole
        description: Change an existing role for a user in an organization.
      tags:
        - web app
      description: Changes an existing role for a user in an organization.
...
  '/user/{userId}/password':
    parameters:
      - $ref: '#/components/parameters/userId'
    put:
      summary: "Update a user's password."
      operationId: put-users-userId-password
      responses:
        '204':
          description: No Content
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/DefaultResponse'
      description: Changes the password of an user.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                oldPassword:
                  type: string
                newPassword:
                  type: string
              required:
                - oldPassword
                - newPassword
      tags:
        - mobile app
...
  /user/password/recovery:
    post:
      description: "Request a link to recover a user's password."
      operationId: post-user-password-recovery
      requestBody:
        content:
          application/json:
            schema:
              properties:
                email:
                  format: email
                  type: string
              required:
                - email
              type: object
      responses:
        '204':
          description: No Content
        default:
          $ref: '#/components/responses/DefaultResponse'
      security: []
      summary: "Create a request to recover a user's password."
      tags:
        - mobile app
...
components:
  parameters:
    userId:
      in: path
      name: userId
      description: The id of a user.
      required: true
      schema:
        $ref: ./model/GenericId.v1.yaml
    organizationId:
      in: path
      name: organizationId
      description: The id of an organization.
      required: true
      schema:
        $ref: ./model/GenericId.v1.yaml
...
  responses:
    BadRequest:
      description: The server could not process the request due to a client mistake.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                enum:
                  - The server could not process the request due to a client mistake.
            required:
              - message
    Unauthorized:
      description: The server could not authenticate the user.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                enum:
                  - The server could not authenticate the user.
            required:
              - message
    Forbidden:
      description: The server could not authorize the user.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                enum:
                  - The server could not authorize the user.
            required:
              - message
    NotFound:
      description: The requested resource does not exist.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                enum:
                  - The requested resource does not exist.
                description: Resource not found.
            required:
              - message
    DefaultResponse:
      description: The server encountered an unexpected error.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                enum:
                  - The server encountered an unexpected error.
                description: Unexpected error.
            required:
              - message
  securitySchemes:
    JWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: The token is sent through the Authorization header with Bearer schema.
security:
  - JWT: []